pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper2epub"
version = "0.1.0"
description = "A powerful academic PDF to EPUB converter with AI-powered layout detection and LaTeX math support"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["pdf", "epub", "converter", "academic", "scientific", "latex", "nougat"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "nougat-ocr>=0.1.17",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "ebooklib>=0.18",
    "Pillow>=10.0.0",
    "markdown>=3.4.0",
    "click>=8.0.0",
    "tqdm>=4.65.0",
    "pymupdf>=1.23.0",
    "albumentations<1.4.0",  # Fix compatibility with nougat-ocr
    "pypdfium2>=4.0.0,<5.0.0",  # Fix PDF rasterization compatibility
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/paper2epub"
Documentation = "https://github.com/yourusername/paper2epub#readme"
Repository = "https://github.com/yourusername/paper2epub"
Issues = "https://github.com/yourusername/paper2epub/issues"

[project.scripts]
paper2epub = "paper2epub.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["paper2epub*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
